refactor(channel): extract serve_runtime.rs from mod.rs
